refactor: tag list 삭제
